summarizer-app/requirements.txt
Flask==2.3.3
google-generativeai==0.8.3
gunicorn==23.0.0
